requirements.txt
altgraph==0.17.4
astroid==2.8.6
attrs==22.1.0
autopep8==1.6.0
azure-common==1.1.28
azure-core==1.27.1
azure-functions==1.17.0
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-datafactory==3.1.0
azure-mgmt-resource==23.0.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
distro==1.8.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.3
mccabe==0.6.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.22.4
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.2
pandas==1.4.2
paramiko==2.11.0
pefile==2023.2.7
platformdirs==2.5.0
portalocker==2.7.0
pycodestyle==2.8.0
pycparser==2.21
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.7.0
pylint==2.11.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
texttable==1.6.4
toml==0.10.2
typing_extensions==4.6.3
urllib3==2.0.3
websocket-client==0.59.0
Werkzeug==2.3.6
wrapt==1.13.3
